Fix overlap between the supress format/diff args
Previously, asking `cfr` to skip formatting the contents of the output
file would also skip producing the diff entirely. While this is the kind
of behaviour you may want to have to deter people from submitting
patches to Linux that do not comply with Rust formatting standards, it
is detrimental to testing (which I understand to be the primary purpose
of suppressor flags).

This patch moves diff-producing code outside of the block guarded by a
check against the key that represents the user's choice to suppress
formatting, allowing CFR to produce a diff even when the processed
output file has not been formatted.

Signed-off-by: Amélie "Lymkwi" Gonzalez <lymkwi@vulpinecitrus.info>