Make rustfmt config path an `Option`
If no path to a `rustfmt` configuration file was provided, this program
would default to `rustfmt.toml`, and force `rustfmt` to try and find it
recursively in all folders, even if it did not end up finding it.

This patch makes the `rustfmt` configuration file path an `Option`,
meaning that no default value is provided. If no file path is given, the
`--config-path` option is never provided to `rustfmt` during its
invocation.

Signed-off-by: Amélie Gonzalez <lymkwi@vulpinecitrus.info>